delete statement nobody will understand
Apple is probably marketing something that is technically incorrect, or, at least, inaccurate. If an activist or a journalist is targeted by a global actor, I would never ever recommend using a classic smartphone for sensitive communications, even if the "extreme mode" is activated.

Most people in that case would use a dedicated system and Tor. However, it's not in the scope of this guide, so there's no need to include such references.

Signed-off-by: J. ⭐⭐ <6878027+jmau111@users.noreply.github.com>